% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_lookup_fn.R
\name{a_lookup_fn}
\alias{a_lookup_fn}
\title{Make lookup table for camera area}
\usage{
a_lookup_fn(x)
}
\arguments{
\item{x}{Camera database with columns cam and a (camera area)}
}
\value{
A dataframe with one row per camera, and its area
}
\description{
This function will create a summary table of all cameras and their areas.
NOTE! Estimates of abundance will be wrong if any cameras are missing from x (i.e., the camera
  was functioning but didn't get any pictures.)
}
\details{
Disclaimer! This function currently does not account for camera areas that change over time.
}
\examples{
df <- data.frame(cam = c(1,1,2,2),
                 datetime = as.POSIXct(c("2016-01-02 12:00:00",
                                         "2016-01-03 13:12:00",
                                         "2016-01-02 14:00:00",
                                         "2016-01-03 16:53:42") ),
                 a = c(850, 850, 1100, 1100))
a_lookup_fn(df)
}
