% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_int.R
\name{add_int}
\alias{add_int}
\title{Add interval column}
\usage{
add_int(x)
}
\arguments{
\item{x}{a dataframe or tibble with columns "start" and "end"}
}
\value{
a dataframe or tibble
}
\description{
Add interval column
}
\examples{
add_int(deploy)
}
