% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_bins.R
\name{assign_bins}
\alias{assign_bins}
\title{Assign each measure of "to_event" to a bin for the Chi squared test}
\usage{
assign_bins(bin_df, to_event)
}
\arguments{
\item{bin_df}{data frame defining the interval of each bin}

\item{to_event}{observations of "to_event"}
}
\value{
Returns the bin that each observation of to_event falls in
}
\description{
Assign each measure of "to_event" to a bin for the Chi squared test
}
\examples{
assign_bins(bin_df, to_event)
}
