% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gof_bins_var.R
\name{build_gof_bins_var}
\alias{build_gof_bins_var}
\title{Define Chi squared bins of variable width at each occasion}
\usage{
build_gof_bins_var(eh, n_bins, lambda)
}
\arguments{
\item{eh}{encounter history object}

\item{n_bins}{integer, the number of bins to create}

\item{lambda}{estimate of density (lambda from poisson and exponential distributions)}
}
\value{
Data frame defining the bin intervals for each occasion
}
\description{
Define Chi squared bins of variable width at each occasion
}
\examples{
build_gof_bins_var(eh, n_bins, lambda)
}
