% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_occ.R
\name{build_occ}
\alias{build_occ}
\title{Build sampling occasions}
\usage{
build_occ(samp_freq, samp_length, study_start, study_end)
}
\arguments{
\item{samp_freq}{numeric, The number of seconds between the start of each sampling occasion}

\item{samp_length}{numeric. The number of seconds to sample at each sampling occasion}

\item{study_start}{POSIXct. The start of the study}

\item{study_end}{POSIXct. The end of the study}
}
\value{
a dataframe
}
\description{
Build sampling occasions
}
\examples{
  study_dates <- as.POSIXct(c("2016-01-01 00:00:00", "2016-01-04 23:59:59"), tz = "GMT")
  build_occ(samp_freq = 3600, 
            samp_length = 10,
            study_start = study_dates[1],
            study_end = study_dates[2])
}
