% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_occ.R
\name{build_occ_cam}
\alias{build_occ_cam}
\title{Sampling occasions by camera}
\usage{
build_occ_cam(deploy, occ)
}
\arguments{
\item{deploy}{deploy object}

\item{occ}{dataframe of sampling occasions}
}
\value{
dataframe
}
\description{
Sampling occasions by camera
}
\examples{
  study_dates <- as.POSIXct(c("2016-01-01 00:00:00", "2016-01-04 23:59:59"), tz = "GMT")
  occ <- build_occ(samp_freq = 3600, 
            samp_length = 10,
            study_start = study_dates[1],
            study_end = study_dates[2])
  build_occ_cam(deploy, occ)
}
