% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_censor.R
\name{calc_censor}
\alias{calc_censor}
\title{Calculate censor for each time step}
\usage{
calc_censor(effort)
}
\arguments{
\item{effort}{dataframe built from effort_fn()}
}
\value{
summarised dataframe with occ and censor columns
}
\description{
Calculate censor for each time step
}
\examples{
 occ <- build_occ(samp_freq = 3600, 
            samp_length = 10,
            study_start = study_dates[1],
            study_end = study_dates[2])
eff <- effort_fn(deploy, occ)
calc_censor(eff)
}
