% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ste.R
\name{calc_ste}
\alias{calc_ste}
\title{Calculate STE on each occasion}
\usage{
calc_ste(df, occ, effort)
}
\arguments{
\item{df}{df object}

\item{occ}{occasions}

\item{effort}{effort from effort_fn()}
}
\description{
Calculate STE on each occasion
}
\examples{
  occ <- build_occ(samp_freq = 3600, 
            samp_length = 10,
            study_start = study_dates[1],
            study_end = study_dates[2])
 eff <- effort_fn(deploy, occ)
 calc_ste(df, occ, eff)
}
