% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chck_names.R
\name{chck_names}
\alias{chck_names}
\title{Check that certain names exist}
\usage{
chck_names(x, colnames)
}
\arguments{
\item{x}{a dataframe}

\item{colnames}{a character vector of column names you want to check exist}
}
\value{
T or F
}
\description{
This isn't different from verify except in its output...and I don't use it anymore...
}
\examples{
chck_names(occ, c("start", "end"))
}
