% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_fn.R
\name{diff_fn}
\alias{diff_fn}
\title{Select rows within an interval}
\usage{
diff_fn(x, y, interval_length)
}
\arguments{
\item{x}{A numeric vector or the name of a column in a dataframe}

\item{y}{A numeric vector of the same length}

\item{interval_length}{The interval between y and x}
}
\value{
A vector of the same length as x, with the first value of y where x-y <= interval_length
}
\description{
Used within mutate, this function creates a new column with the first value of y
where x-y is less than or equal to a specified interval.
}
\examples{
df <- data.frame(ID = 1:4, Value = c(1, 5, 10, 8))
y <- c(2,4,6,8)
dplyr::mutate(df, newcol = diff_fn(Value, y, 1))

}
