% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_density.R
\name{exp_dens}
\alias{exp_dens}
\title{Calculate exponential density between two values}
\usage{
exp_dens(left, right, lambda)
}
\arguments{
\item{left}{leftmost values}

\item{right}{righmost values}

\item{lambda}{lambda, from the exponential distribution}
}
\value{
returns value between 0 and 1
}
\description{
Calculate exponential density between two values
}
\examples{
exp_dens(0.1, 0.2, 1)
}
