% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_gof_test.R
\name{exp_gof_test}
\alias{exp_gof_test}
\title{Title Background function; single exponential goodness of fit test}
\usage{
exp_gof_test(eh, n_bins, lambda, bin_cuts = "fixed")
}
\arguments{
\item{eh}{eh df}

\item{n_bins}{number of bins to use in chi squared test}

\item{lambda}{estiamte of lambda}

\item{bin_cuts}{"fixed" or "variable"}
}
\value{
goodness of fit result
}
\description{
Title Background function; single exponential goodness of fit test
}
\examples{
exp_gof_test(eh, n_bins, lambda, bin_cuts = "fixed")
}
