% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_gof_TTE.R
\name{exp_gof_tte}
\alias{exp_gof_tte}
\title{Exponential goodness-of-fit tests for TTE}
\usage{
exp_gof_tte(n_bins, lambda, eh, bin_cuts = "fixed")
}
\arguments{
\item{n_bins}{a vector of the number of bins to use in the chi squared tests}

\item{lambda}{Estimate of density calculated from the output of tte_estn_fn}

\item{eh}{encounter history from tte_build_eh}

\item{bin_cuts}{"fixed" or "variable". Fixed bins have the same cutoff values for every occasion. Variable bins allow the cutoff values to change to keep the bin widths equal.}
}
\value{
Returns a summary table showing the results of all of the chi squared tests and the chi squared table for each test.
}
\description{
Exponential goodness-of-fit tests for TTE
}
\examples{
exp_gof_TTE(eh, n_bins, lambda, bin_cuts = "fixed")
}
