% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_logl_fn.R
\name{exp_logl_fn}
\alias{exp_logl_fn}
\title{Exponential log likelihood}
\usage{
exp_logl_fn(x, param)
}
\arguments{
\item{x}{A list formulated by ste_data_fn or tte_data_fn. 
x$toevent is a matrix with space- or time-to-event and NAs.
x$censor is the censor}

\item{param}{The value of log(lambda) of which you want to evaluate the logL.}
}
\value{
The MLE value of lambda-hat
}
\description{
The log likelihood of the exponential distribution that is used for STE and TTE.
}
\details{
For TTE, x$toevent will be a matrix with 1 row per camera and 1 column per sampling occasion. 
For STE, x$toevent will be a matrix with 1 row. 1 column per sampling occasion
}
\examples{
x <- build_ste_eh(df, deploy, occ)
dat.ste <- list(toevent = matrix(x$STE, nrow = 1),
                censor = x$censor)
optim(log(1/mean(dat.ste$toevent, na.rm = T)), 
             exp_logl_fn, 
             x = dat.ste, 
             control = list(fnscale = -1),
             hessian = T)
  
}
