% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{find_overlap}
\alias{find_overlap}
\title{Find cameras with overlapping intervals}
\usage{
find_overlap(x)
}
\arguments{
\item{x}{a dataframe or tibble, with columns cam, start and end}
}
\value{
a dataframe with any records that are overlapping (grouped by cam)
}
\description{
Find cameras with overlapping intervals
}
