% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_bins_var_int.R
\name{gof_bins_var_int}
\alias{gof_bins_var_int}
\title{Internal function to build gof bins at a single occasion}
\usage{
gof_bins_var_int(censor, lambda, n_bins)
}
\arguments{
\item{censor}{right-censor cutoff}

\item{lambda}{estimate of lambda}

\item{n_bins}{number of bins to build}
}
\value{
returns a one row data frame
}
\description{
Internal function to build gof bins at a single occasion
}
\examples{
gof_bins_var_int(censor, lambda, n_bins)
}
