% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ise_estN_fn.R
\name{ise_estN_fn}
\alias{ise_estN_fn}
\title{Estimate N through ISE}
\usage{
ise_estN_fn(ise_eh, study_area)
}
\arguments{
\item{ise_eh}{an instantaneous sampling encounter history}

\item{study_area}{the size of the study area (same units as camera area)}
}
\value{
a tibble with density and abundance estimate
}
\description{
Estimate N through ISE
}
\examples{
ise_eh <- ise_build_eh(df, deploy, occ) 
ise_estN_fn(ise_eh, 1e6)
}
