% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_facet.R
\name{label_facet}
\alias{label_facet}
\title{Custom function to rename the facets}
\usage{
label_facet(original_var, custom_name)
}
\arguments{
\item{original_var}{Column name of the facet label in the ggplot object}

\item{custom_name}{Characters to add to before the current facet label}
}
\value{
A character vector
}
\description{
Custom function to rename the facets
}
\examples{
label_facet(rslt$TrueN, "N")
}
