% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCI.R
\name{logCI}
\alias{logCI}
\title{Log confidence intervals}
\usage{
logCI(estimate, se)
}
\arguments{
\item{estimate}{parameter estimate}

\item{se}{estimate parameter standard error}
}
\value{
a data frame with LCI and UCI
}
\description{
Returns confidence intervals for abundance that do not go below 0
}
\examples{
logCI(32, 25)
}
