% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_pos.R
\name{only_pos}
\alias{only_pos}
\title{Replace negative values with NA}
\usage{
only_pos(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of the same length as x
}
\description{
Retains 0 and positive values, and replaces negative values with NA
}
\examples{
x <- seq(-2, 2, by = 1)
only_pos(x)
}
