% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overl}
\alias{overl}
\title{Find if a single interval overlaps multiple intervals}
\usage{
overl(one_int, n_int)
}
\arguments{
\item{one_int}{a single lubridate::interval}

\item{n_int}{a vector of lubridate::intervals}
}
\value{
a single value, the number of intervals that one_int overlaps
}
\description{
Find if a single interval overlaps multiple intervals
}
\examples{
s <- Sys.time()
a <- lubridate::interval((s-5), (s))
b <- lubridate::int_diff(c((s-4),(s+1),(s+5)))
overl(a, b)
}
