% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwr_cv.R
\name{plot_pwr_cv}
\alias{plot_pwr_cv}
\title{Plot coefficient of variation from power analysis simulations}
\usage{
plot_pwr_cv(pwr_sim_rslt)
}
\arguments{
\item{pwr_sim_rslt}{Result from ste_sim_fn or tte_sim_fn}
}
\value{
a ggplot
}
\description{
Plot coefficient of variation from power analysis simulations
}
\examples{
rslt <- ste_pwr_sim(
  N = c(1, 2),
  study_area = c(1e6, 1e8),
  ncam = c(30, 50),
  nocc = c(1000, 2000),
  cam_area = c(100, 150),
  niter = 5
)
plot_pwr_cv(rslt)
}
