% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwr_sim.R
\name{plot_pwr_sim}
\alias{plot_pwr_sim}
\title{Plot each iteration from power analysis simulation}
\usage{
plot_pwr_sim(pwr_sim_rslt)
}
\arguments{
\item{pwr_sim_rslt}{Result from ste_sim_fn or tte_sim_fn}
}
\value{
a ggplot
}
\description{
Plot each iteration from power analysis simulation
}
\examples{
rslt <- ste_pwr_sim(
  N = c(1, 2),
  study_area = c(1e6, 1e8),
  ncam = c(30, 50),
  nocc = c(1000, 2000),
  cam_area = c(100, 150),
  niter = 5
)
plot_pwr_sim(rslt)
}
