% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_density.R
\name{pois_dens}
\alias{pois_dens}
\title{Poisson density function}
\usage{
pois_dens(x, lambda)
}
\arguments{
\item{x}{count}

\item{lambda}{parameter of the poisson distribution}
}
\value{
returns the probability of count x
}
\description{
Poisson density function
}
\examples{
pois_dens(3,1)
}
