% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gof_STE.R
\name{pois_gof_ste}
\alias{pois_gof_ste}
\title{Poisson goondess-of-fit tests for STE}
\usage{
pois_gof_ste(n_bins, lambda, df, deploy, occ)
}
\arguments{
\item{n_bins}{A vector of the number of bins to use in the chi squared test}

\item{lambda}{Calculated from the output of ste_estn_fn; EstN}

\item{df}{data frame prepped for ste_build_eh}

\item{deploy}{deployment data frame prepped for ste_build_eh}

\item{occ}{occasion data frame prepped for ste_build_eh}
}
\value{
returns a summary of the goodness-of-fit tests and the chi_squared results for each test
}
\description{
Poisson goondess-of-fit tests for STE
}
\examples{
pois_gof_STE(n_bins, lambda, df, deploy, occ)
}
