% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gof_test.R
\name{pois_gof_test}
\alias{pois_gof_test}
\title{Single iteration of the poisson goodness-of-fit test}
\usage{
pois_gof_test(n_bins, count, lambda)
}
\arguments{
\item{n_bins}{number of bins to use in chi-squared test}

\item{count}{vector of observed counts}

\item{lambda}{estimate of lambda}
}
\value{
returns the result of the goodness-of-fit test
}
\description{
Single iteration of the poisson goodness-of-fit test
}
\examples{
pois_gof_test(n_bins, count, lambda)
}
