% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gof_TTE.R
\name{pois_gof_TTE}
\alias{pois_gof_TTE}
\title{Poisson goodness-of-fit tests for TTE}
\usage{
pois_gof_TTE(n_bins, lambda, eh)
}
\arguments{
\item{n_bins}{a vector of the number of bins to use in each chi squared test}

\item{lambda}{estimate of density calculated from the tte_estN_fn output}

\item{eh}{encounter history from tte_build_eh}
}
\value{
Returns a summary table for the tests and the chi squared table from each test
}
\description{
Poisson goodness-of-fit tests for TTE
}
\examples{
pois_gof_TTE(n_bins, lambda, eh)
}
