% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ste_sim_fn.R
\name{ste_sim_fn}
\alias{ste_sim_fn}
\title{STE simulation for one combination of values}
\usage{
ste_sim_fn(N, study_area, ncam, nocc, cam_area, pb)
}
\arguments{
\item{N}{a single value, true abundance for simulation}

\item{study_area}{a single value, true study area size (N/study_area = true density)}

\item{ncam}{a single value, number of cameras to simulate}

\item{nocc}{a single value, number of occasions to simulate}

\item{cam_area}{a single value, size of the camera viewshed}
}
\value{
a tibble
}
\description{
STE simulation for one combination of values
}
\examples{
ste_pwr_sim(
  N = c(5, 15)
  study_area = 1e6,
  ncam = c(50, 100),
  nocc = c(10000, 20000),
  cam_area = 150,
  niter = 30
)
}
