% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_build_occ.R
\name{tte_build_occ}
\alias{tte_build_occ}
\title{tte_build_occ}
\usage{
tte_build_occ(per_length, nper, time_btw, study_start, study_end)
}
\arguments{
\item{per_length}{numeric. Length of the TTE sampling period.}

\item{nper}{numeric. Number of TTE sampling periods per sampling occasion.}

\item{time_btw}{numeric. Length of time between sampling occasions, 
allowing animals to re-randomize.}

\item{study_start}{POSIXct. The start of the study.}

\item{study_end}{POSIXct. The end of the study.}
}
\value{
a dataframe
}
\description{
tte_build_occ
}
\examples{
deploy <- data.frame(
  cam = c(1, 2, 2, 2),
  start = as.POSIXct(c("2015-12-01 15:00:00",
                       "2015-12-08 00:00:00", 
                       "2016-01-01 00:00:00", 
                       "2016-01-02 00:00:00"),
                     tz = "GMT"),
  end = as.POSIXct(c("2016-01-05 00:00:00", 
                     "2015-12-19 03:30:00", 
                     "2016-01-01 05:00:00",
                     "2016-01-05 00:00:00"), 
                   tz = "GMT"),
  area = c(300, 200, 200, 450)
)
per <- tte_samp_per(deploy, lps = 30/3600)
study_dates <- as.POSIXct(c("2016-01-01 00:00:00", "2016-01-04 23:59:59"), tz = "GMT")
occ2 <- tte_build_occ(
  per_length = per,
  nper = 24,
  time_btw = 2 * 3600,
  study_start = study_dates[1],
  study_end = study_dates[2]
)
}
