% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_data_fn.R
\name{tte_data_fn}
\alias{tte_data_fn}
\title{Create TTE encounter history}
\usage{
tte_data_fn(x, countcol, speed, j, samp_freq, cam_areas, date_lim)
}
\arguments{
\item{x}{A dataframe with columns named cam (any class) and datetime (class POSIXct)
and a column (any name) of the count of target species in each photo.}

\item{speed}{Average speed of the study animal, in units/sec. Match units to a and A!}

\item{j}{}

\item{samp_freq}{how often you want to begin sampling, in seconds (e.g. 
3600 means sample every hour).}

\item{cam_areas}{A dataframe made by a_lookup_fn or similar (with columns cam and a).
Must have one row per active camera and that camera's area (a)}

\item{date_lim}{A vector of length 2 of class POSIXct. The first and last date of the
desired sampling period.}

\item{count_col}{The name of the column containing the count of animals}
}
\value{
A list with the encounter history for a Time-to-Event model.
}
\description{
At every camera i and sampling occasion j, find the time-to-event.
}
\details{
Disclaimer! This function cannot currently account for cameras with different viewable areas.
}
\examples{
df <- data.frame(cam = c(1,1,2,2),
            datetime = as.POSIXct(c("2016-01-02 12:00:00",
                                    "2016-01-03 13:12:00",
                                    "2016-01-02 14:00:00",
                                    "2016-01-03 16:53:42"),
                                   tz = "GMT"),
                  a = c(850, 850, 1100, 1100),
                  count = c(1, 0, 0, 2))
tab <- a_lookup_fn(df)
d <- as.POSIXct(c("2016-01-01 00:00:00", "2016-01-04 23:59:59"), tz = "GMT")
ste_data_fn(df,
            count_col = "count",
            speed = 108000,
            j = 4,
            samp = 3600,
            cam_areas = tab,
            date_lim = d,
            A = 150000)
}
