% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_pwr_sim.R
\name{tte_pwr_sim}
\alias{tte_pwr_sim}
\title{Simulation-based TTE power analysis}
\usage{
tte_pwr_sim(N, study_area, ncam, nocc, nper, cam_area, niter)
}
\arguments{
\item{N}{a numeric vector, true abundance for simulation}

\item{study_area}{a numeric vector, true study area size (N/study_area = true density)}

\item{ncam}{a numeric vector, number of cameras to simulate}

\item{nocc}{a numeric vector, number of occasions to simulate}

\item{nper}{a numeric vector, number of sampling periods per sampling occasion}

\item{cam_area}{a numeric vector, size of the camera viewshed}

\item{niter}{a single value, number of simulation iterations for each N x 
study_area x ncam x nocc combination}
}
\value{
a tibble
}
\description{
Simulation-based TTE power analysis
}
\examples{
tte_pwr_sim(
  N = c(5, 15),
  study_area = 1e6,
  ncam = c(50, 100),
  nocc = c(10000, 20000),
  nper = 20
  cam_area = 150,
  niter = 30
)
}
