% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_sim_fn.R
\name{tte_sim_fn}
\alias{tte_sim_fn}
\title{TTE simulation for one combination of values}
\usage{
tte_sim_fn(N, study_area, ncam, nocc, nper, cam_area, pb)
}
\arguments{
\item{N}{a single value, true abundance for simulation}

\item{study_area}{a single value, true study area size (N/study_area = true density)}

\item{ncam}{a single value, number of cameras to simulate}

\item{nocc}{a single value, number of occasions to simulate}

\item{nper}{a numeric vector, number of sampling periods per sampling occasion}

\item{cam_area}{a single value, size of the camera viewshed}

\item{pb}{progress bar}
}
\value{
a tibble
}
\description{
TTE simulation for one combination of values
}
\examples{
tte_pwr_sim(
  N = c(5, 15),
  study_area = 1e6,
  ncam = c(50, 100),
  nocc = c(10000, 20000),
  nper = 20
  cam_area = 150,
  niter = 30
)
}
