% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_deploy}
\alias{validate_deploy}
\title{Validate deploy object}
\usage{
validate_deploy(deploy)
}
\arguments{
\item{deploy}{deploy object}
}
\value{
deploy if it passes all tests, error otherwise
}
\description{
Validate deploy object
}
\examples{
validate_deploy(deploy)
}
