% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_df}
\alias{validate_df}
\title{Validate df}
\usage{
validate_df(df)
}
\arguments{
\item{df}{object df}
}
\value{
df if it passes all tests. Otherwise returns error
}
\description{
Make sure df fits all requirements
}
\examples{
validate_df(df)
}
