% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_df_deploy}
\alias{validate_df_deploy}
\title{Validate df and deploy together}
\usage{
validate_df_deploy(df, deploy)
}
\arguments{
\item{df}{df object}

\item{deploy}{deploy object}
}
\value{
error if any check fails, nothing otherwise
}
\description{
Validate df and deploy together
}
\examples{
validate_df_deploy(df, deploy)
}
