% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_occ}
\alias{validate_occ}
\title{Validate occ}
\usage{
validate_occ(occ)
}
\arguments{
\item{occ}{occ object}
}
\value{
occ if passes all tests, error otherwise
}
\description{
Validate occ
}
\examples{
validate_occ(occ)
}
