% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_start_end}
\alias{validate_start_end}
\title{Validate start and end columns}
\usage{
validate_start_end(x)
}
\arguments{
\item{x}{a dataframe or tibble with columns start and end}
}
\value{
x if it passes all tests, an error otherwise
}
\description{
Validate any object with start and end columns
}
\examples{
validate_start_end(deploy)
}
