% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_gof_STE.R
\name{exp_gof_STE}
\alias{exp_gof_STE}
\title{Exponential goodness-of-fit tests for STE}
\usage{
exp_gof_STE(eh, n_bins, lambda = NULL, bin_cuts = "fixed")
}
\arguments{
\item{eh}{encounter history df with the columns required by spaceNtime}

\item{n_bins}{the number of bins to use in the chi squared test}

\item{lambda}{estimate of lambda calculated from the ste_estn_fn output}

\item{bin_cuts}{"fixed" or "variable". Fixed bins use the same cutoff for each occasion. Variable bins use different cutoffs but keep the width of bins equal within occasions.}
}
\value{
Returns a summary table of all goodness of fit tests and the chi squared results for each test
}
\description{
Exponential goodness-of-fit tests for STE
}
\examples{
exp_gof_STE(eh, n_bins, lambda = NULL, bin_cuts = "fixed")
}
