% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ste_estn_fn.R
\name{ste_estN_fn}
\alias{ste_estN_fn}
\title{Title}
\usage{
ste_estN_fn(x)
}
\arguments{
\item{x}{A list formulated by ste_data_fn or tte_data_fn. 
x$toevent is a matrix with space- or time-to-event and NAs.
x$censor is the censor
x$A is the size of the study area, in the same units as a}
}
\value{
A list, with the estimated abundance with its standard error and confidence intervals.
}
\description{
Title
}
\examples{
df <- data.frame(cam = c(1,1,2,2),
            datetime = as.POSIXct(c("2016-01-02 12:00:00",
                                    "2016-01-03 13:12:00",
                                    "2016-01-02 14:00:00",
                                    "2016-01-03 16:53:42"),
                                   tz = "GMT"),
                  a = c(850, 850, 1100, 1100),
                  count = c(1, 0, 0, 2))
tab <- a_lookup_fn(df)
d <- as.POSIXct(c("2016-01-01 00:00:00", "2016-01-04 23:59:59"), tz = "GMT")
dat.ste <- ste_data_fn(df,
            countcol = "count",
            samp = 3600,
            samplength = 10,
            camareas = tab,
            datelim = d,
            A = 150000)
ste_estN_fn(dat.ste)
}
